requirements.txt
requests>=2.31.0,<3.0.0
python-dateutil>=2.8.2,<3.0.0
tqdm>=4.66.0,<5.0.0
